+++ CHANGELOG.md
@@ -63,6 +63,8 @@
   - Small adjustments to `LongTimeTrace` and tests. (!1014)
   - Make it possible to override gate operations with device level parameters. (!1018)
   - New schedule `cnot_as_h_cz_h` in composite_factory. (!1020)
+- QuantumDevice
+  - Introduce BasicSpinElement spin device element. (!1015)
 
 ## 0.20.1 (2024-05-01)